feat: implement the recursive composition finding in hct-components script
Add a loop that iterates over every composition option returned from the hct-composition call.
Add a loop that iterates over every character of the passed composition.
The second loop calls the get_character_components (where the loop is located) creating the recursive feature of the script.
The recursion is broken when a character can't be decompossed further, i.e. when it is a single stroke.
All the valid components options are put into an array.
For now, the first valid option is simply used as the chosen one.