refactor: change the way the composition array is created in hct-components script
A separate array with the source information is not needed as source-related filtering
is done within the call to the 'composition' command already.
This allows to print the full composition in verbose mode, i.e. including the IDCs and source info.
The IDCs and source info are now removed only after printing the verbose output within the for loop.