feat: implement the 'tiebreaker' functionality in hct-components script
If there is only one option to choose from, simply choose it.
If there is more than one option, chose the final one according to the chosen rule,
either the first available option or the shortest one.
The 'lenght' rule, i.e. where the shortest option is chosen, is not implemented yet.