feat: allow to pass files as input for hct-components script
The output of the script now depends on whether the input is a single character or a file.
If the input is a file, the progress status is printed to the stderr stream.
If input is a single character and and error is met, also print an error message to the stderr stream.